requirements.txt
attrs==19.3.0
certifi==2020.4.5.1
chardet==3.0.4
click==7.1.2
dataclasses==0.7
fastapi==0.54.1
h11==0.9.0
httptools==0.1.1
idna==2.9
importlib-metadata==1.6.0
more-itertools==8.2.0
packaging==20.3
Pillow==7.1.2
pluggy==0.13.1
py==1.8.1
pydantic==1.5.1
pyparsing==2.4.7
pytest==5.4.1
redis==3.5.0
requests==2.23.0
rq==1.3.0
six==1.14.0
starlette==0.13.4
urllib3==1.25.9
uvicorn==0.11.5
uvloop==0.14.0
wcwidth==0.1.9
websockets==8.1
zipp==3.1.0
